requirements.txt
langgraph>=0.0.52
langchain-core>=0.1.50
langchain-openai>=0.0.10
openai>=1.14.0
streamlit>=1.32.0
sqlalchemy>=2.0.30
pydantic>=2.7.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
requests>=2.31.0
networkx>=3.2.1
plotly>=5.18.0
json5>=0.9.14
rich>=13.7.0
tenacity>=9.0.0
flask>=3.0.0